+++ src/Curiosity.EMail.Smtp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.2.0] - 2021-12-02
+
+### Changed
+
+- `IEmailLogger` returns `Response` class object.
+
 ## [1.1.0] - 2021-09-30
 
 ### Added